+++ CHANGELOG.md
@@ -5,6 +5,7 @@ changes with their rationale when appropriate:
 
 ### v5.28.2.0
 - **http4k-*** : Upgrade some dependency versions including Kotlin to 2.0.20
+- **http4k-server-*** : [Fix #1163] Request.version is always HTTP/1.1 for server requests, even when using HTTP/2.0
 - **http4k-*** : [New module] Azure SDK integration - use an http4k client with the official libraries
 
 ### v5.28.1.0